Upgrade to Spring Boot 1.4
Let's upgrade our application to 1.4. We can see we have some warnings
now that we'll have to address. As of Spring Boot 1.4, AssertJ is
provided by `spring-boot-starter-test` so we can remove the duplicate.
Spring Boot 1.4 also provides dependency management for EhCache so we
can remove our dependency management now too.