Migrate Controller to WebMvcTest
Rather than starting our entire application as a web application and
configuring `MockMvc` ourselves we use the new "test slicing" feature
of Spring Boot 1.4

When using the dedicated `WebMvcTest` annotation, we tell Spring Boot
we want to test a controller and that it has to do several things for us
such as auto-configuring `MockMvc`.

Another new feature of Spring Boot 1.4 demoed in this commit is to use
`MockBean` rather than managing the mock ourselves.

Slicing will tune component scan to only enable very specific beans. Our
custom security config isn't applied anymore and we're back with the
default auto-config that doesn't work for us since `/submit` is public.
We import our `SecurityConfig` to make sure security is applied for this
test.

Also we can remove our override of `@EnableTransactionManagement`. Spring
Boot 1.4 set this value to `true` by default but even if we set it to
`false`, the infrastructure behind `MockBean` takes care of that problem
for us.