Migrate rest client test to RestClientTest
So far, `GithubClient` has a protected constructor to be able to pass
the `RestTemplate` to use for testing purposes. This is important as the
`RestTemplate` that the class to test and the one that
`MockRestServiceServer` use must be the same.

Spring Boot 1.4 introduces `RestTemplateBuilder`, a facility to easily
construct a `RestTemplate` with additional customizations. Rather
than having this protected constructor, we now inject the
`RestTemplateBuilder` in our main constructor and that allows us to
use `RestClientTest` as well. This specialized test annotation auto
configures `MockRestServiceServer` for us so all we have to do is inject
it.

Because we use the Spring runner, we can also automatically create a mock
for the counter service.