Add Integration test
It is of course still possible to start a full context as before with the
new `@SpringBootTest` annotation. In this commit we add another
variant of the `CfpController` test that actually exercises the whole
application.

We still want to use `MockMvc` though so we use `@AutoconfigureMockMvc`
to enable that. We could have tuned the `webEnvironment` parameter to
actually start the web container. In this case, a `TestRestTemplate` bean
is automatically made available.